+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.0
 
-* API update: new `launch` method with just one API key argument for both platforms; remove old `launch` method; remove all old `launch...` methods.
+* API update: new `launch` method with just one API key argument for both platforms. Contact us at hi@qonversion.io to merge your old API keys into one, [we can do it now](https://qonversion.io/docs/crossplatform-project).
+* API update: Remove old `launch` method; remove all old `launch...` methods.
 * Add `trackPurchase` method to track Android purchases manually.
 * Add `addAttributionData` method implementation for Android.